requirements/base.txt
Django==3.2.7
djangorestframework==3.13.1
python-dotenv
psycopg2-binary
python-nmap
drf-yasg
djangorestframework-simplejwt
urllib3
requests
